data/requirements.txt
weaviate-client
python-dotenv
firecrawl-py
groq